+++ modules/cloud_run_service/main.tf
@@ -54,6 +54,7 @@ resource "google_cloud_run_service" "default" {
   traffic {
     percent         = 100
     latest_revision = true
+    tag             = var.revision_tag
   }
 
   metadata {